+++ crates/affinidi-messaging/affinidi-messaging-mediator/src/handlers/authenticate.rs
@@ -16,10 +16,7 @@ use affinidi_messaging_didcomm::{Message, UnpackOptions, envelope::MetaEnvelope}
 use affinidi_messaging_mediator_common::errors::{AppError, MediatorError, SuccessResponse};
 use affinidi_messaging_sdk::{
     messages::{AuthorizationResponse, GenericDataStruct, known::MessageType},
-    protocols::mediator::{
-        accounts::AccountType,
-        acls::{AccessListModeType, MediatorACLSet},
-    },
+    protocols::mediator::{accounts::AccountType, acls::MediatorACLSet},
 };
 use axum::{Json, extract::State};
 use http::StatusCode;
@@ -82,30 +79,23 @@ pub async fn authentication_challenge(
         // 2. If not known, then does the mediator acl_mode allow for new accounts?
         // 3. If yes, then add the account and continue
 
-        match state.database.get_did_acl(&session.did_hash).await? {
-            Some(acls) => {
-                if acls.get_blocked() {
-                    info!("DID({}) is blocked from connecting", session.did);
-                    return Err(MediatorError::ACLDenied("DID Blocked".to_string()).into());
-                }
-            }
-            _ => {
-                // Unknown DID
-                if state.config.security.mediator_acl_mode == AccessListModeType::ExplicitAllow {
-                    info!("Unknown DID({}) is blocked from connecting", session.did);
-                    return Err(MediatorError::ACLDenied("DID Blocked".to_string()).into());
-                } else {
-                    // Register the DID as a local DID
-                    state
-                        .database
-                        .account_add(
-                            &session.did_hash,
-                            &state.config.security.global_acl_default,
-                            None,
-                        )
-                        .await?;
-                }
-            }
+        // Check if DID is allowed to connect
+        let (allowed, known) =
+            MediatorACLSet::authentication_check(&state, &digest(&session.did_hash), None).await?;
+
+        if !allowed {
+            info!("DID({}) is blocked from connecting", session.did);
+            return Err(MediatorError::ACLDenied("DID Blocked".to_string()).into());
+        } else if !known {
+            // Register the DID as a local DID
+            state
+                .database
+                .account_add(
+                    &session.did_hash,
+                    &state.config.security.global_acl_default,
+                    None,
+                )
+                .await?;
         }
 
         state.database.create_session(&session).await?;
@@ -163,7 +153,10 @@ pub async fn authentication_response(
         let from_did = match &envelope.from_did {
             Some(from_did) => {
                 // Check if DID is allowed to connect
-                if !MediatorACLSet::authentication_check(&state, &digest(from_did), None).await? {
+                if !MediatorACLSet::authentication_check(&state, &digest(from_did), None)
+                    .await?
+                    .0
+                {
                     info!("DID({}) is blocked from connecting", from_did);
                     return Err(MediatorError::ACLDenied("DID Blocked".to_string()).into());
                 }